requirements.txt
absl-py==0.12.0
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
cycler==0.10.0
decorator==5.0.9
dill==0.3.4
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.34.1
h5py==3.1.0
idna==3.2
importlib-metadata==4.6.3
importlib-resources==5.2.2
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.4.2
matplotlib-inline==0.1.2
MouseInfo==0.1.3
mss==6.1.0
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
pandas==1.1.5
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
promise==2.3
prompt-toolkit==3.0.19
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
Pygments==2.9.0
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyRect==0.1.4
PyScreeze==0.1.27
pytesseract==0.3.8
python-dateutil==2.8.2
PyTweening==1.0.3
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.5.0
tensorflow-metadata==1.2.0
termcolor==1.1.0
tqdm==4.62.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
